Shortcut for "Test" instructions; fixed instruction names
Earlier I built a shortcut for making new Math instructions;
test instructions are similar, but write to a different stack.
Now it's simple to make new tests as well.

Instruction names used to be limited to the actual name of
the instruction and thats it (unless there was a specifier for
the instruction like Nop:0).  Now all instructions can have a
':' and any aditional characters to give them a unique name.

For example, "Inc" and "Inc:v2" would operate identically, but
technically have different names.