Added automatic mechanisms to deal with return values
When we run many functions at once, it can be annoying to have to deal with all of
the return values.  Three new methods will help with that: FindMax, FindMin, and FindSum.

These will run all of the stored functions and return the appropriate value.

Note that for Boolean return types, FindMax can be used to see if ANY are true, and FindMin
to see if any are false.