More work on bit matricies
It looks like the to_ullong method for bitset doesn't just grab the last 64 bits,
but throws and exception if the bitset is too big.  This is a bit frustrating for
working with bitsets.

I'll probably shift back to creating my own, but focus on previous implementations
of bitsets in STL.