Added implementation of attrs from plotting work.
The attrs.h file contains the implementation of attributes and attribute
packs, which allows the user to create structs (or attribute packs) on the fly at
compile-time from pre-defined attributes.

The flow.h file contains some utilities which allow the construction of
dataflows, which are a special feature of attribute packs and other
objects allowing them to be used as maps.