Make PrintHeaderKeys optional in DataFile builders
If set false, the added optional boolean argument (default true) for
SetupFitnessFile, SetupSystematicsFile, and SetupPopulationFile
suppresses the call to PrintHeaderKeys in these methods.
This allows for new columns to be added to the default Systematics,
Fitness, and Population DataFile configuration. The user is then
responsible for calling PrintHeaderKeys after adding the desired
columns have been added. This is especially useful for adding a column
with a unique identifying value for a particular replicate (e.g. seed
value), although I can imagine many other fields that a user may want
to add to these files.