Use Nullstream for emscripten
Currently, DataFile output sent to an ofstream gets sent to memory
which can quickly lead to crashes in the browser due to excessive
memory use. Instead, when compiling with enscripten let's use the
NullStream class instead of the ofstream class. This way, no
operation is performed when the stream is written to and excessive
memory crashes in the browser are a thing of the past.