Add Inst_RngDouble to EventDrivenGP.h
This instruction draws a value between 0 and 1 from the hardware's onboard RNG and stores it in Local[Arg1]. It's potentially useful to many users, so worthwhile to have available in the header files. Due to political considerations, though, don't add it to the default instruction set so that, by default, SignalGP programs are deterministic.