Parse command name to _command, not _positional
Currently, the command name (e.g., the zeroth argument) will be
treated as a positional argument. Let's do something smarter and
assign it to its own key --- "_command". The default argument
handlers (callbacks or ProcessBuiltins) will just throw it away.
However, if the user wants to access it before then it's there!