Revert "Grow by approx golden ratio, not doubling"
This reverts commit c03abe1b38eae0b134b6809755c293213be87551.

There was a bug with size 1 buffers not growing at all
(e.g., 3/2 * 1 = 1).
This caused some nasty segfaults.

However, even with this fixed the class still had incorrect
behavior (failing tests). Not sure why, but I suspect this might
have to do with some of the opaque internal assumptions of
streambuf.