Don't condition providing ids on having content
This was creating a chicken and the egg bug. You want to put something
inside of a table? The table cell needs to have an ID. But the table
cell won't have an ID unless there's something inside of it.

Just give everything an ID always -- there's no reason not to.