Implemented GetBytes method that returns a span of const field_t
A span is to an array what a string_view is to a string. Basically, it
is a lightweight and fast non-owning way to pass arrays
around while conserving their length.
Right now, we are using span-lite, an implementation of std::span, since
the later one will be available in C++20